refactor(Address): convert Address in String type to Address type to split table
