+++ src/main/java/personal/delivery/member/MemberController.java
@@ -17,24 +17,23 @@ import personal.delivery.member.service.MemberService;
 public class MemberController {
 
     private final MemberService memberService;
-//    private final PasswordEncoder passwordEncoder;
 
     @PostMapping("sellers")
     public ResponseEntity<MemberResponseDto> signUpSeller(@RequestBody MemberDto memberDto) {
 
-//        MemberResponseDto memberResponseDto = memberService.createMember(memberDto, passwordEncoder);
         MemberResponseDto memberResponseDto = memberService.createSeller(memberDto);
 
         return ResponseEntity.status(HttpStatus.OK).body(memberResponseDto);
+
     }
 
     @PostMapping("customers")
     public ResponseEntity<MemberResponseDto> signUpCustomer(@RequestBody MemberDto memberDto) {
 
-//        MemberResponseDto memberResponseDto = memberService.createMember(memberDto, passwordEncoder);
         MemberResponseDto memberResponseDto = memberService.createCustomer(memberDto);
 
         return ResponseEntity.status(HttpStatus.OK).body(memberResponseDto);
+
     }
 
 }
