+++ src/main/java/personal/delivery/member/domain/Address.java
@@ -1,0 +1,27 @@
+package personal.delivery.member.domain;
+
+import jakarta.persistence.Embeddable;
+import lombok.Getter;
+
+@Embeddable
+@Getter
+public class Address {
+
+    private String city;
+    private String street;
+    private String zipcode;
+    private String detailedAddress;
+
+    protected Address() {
+    }
+
+    public Address(String city, String street, String zipcode, String detailedAddress) {
+
+        this.city = city;
+        this.street = street;
+        this.zipcode = zipcode;
+        this.detailedAddress = detailedAddress;
+
+    }
+
+}
