+++ src/main/java/personal/delivery/member/dto/MemberDto.java
@@ -1,6 +1,7 @@
 package personal.delivery.member.dto;
 
 import lombok.Data;
+import personal.delivery.member.domain.Address;
 
 @Data
 public class MemberDto {
@@ -8,6 +9,6 @@ public class MemberDto {
     private String name;
     private String email;
     private String password;
-    private String address;
+    private Address address;
 
 }
